\name{HelloGoodbye}
\alias{HelloGoodbye}

\title{Anonymous Binary Data}
\description{
  Real data contributed from an anonymous organization. We borrowed the wording of a famous song to hide the true nature of the data.
}
\usage{
data(HelloGoodbye)
}
\format{
  A data frame of with 13954 observations (rows) and 56 variables (columns).
}
\source{
  An anonymous organization.
}
\keyword{datasets}
