\name{JNcharacter}
\alias{JNcharacter}

\title{The Japanese National Character}
\description{
  A part of the Survey on the Japanese National Character.
}
\usage{
data(JNcharacter)
}
\format{
  A data frame with 85 observations on the following 10 variables.

  A part of the source data was deleted and recoded according to an input
  example of original program CATDAP-01.

  \tabular{rlll}{
  [, 1] \tab sex \tab\tab 1, 2\cr
  [, 2] \tab age \tab\tab 1, 2, 3, 4\cr
  [, 3] \tab pol.party \tab\tab 1, 2, 3, 4\cr
  [, 4] \tab education \tab\tab 1, 2, 3\cr
  [, 5] \tab occupation \tab\tab 1, 2\cr
  [, 6] \tab born.again \tab\tab 1, 2\cr
  [, 7] \tab difficult \tab\tab 1, 2\cr
  [, 8] \tab pleasure \tab\tab 1, 2\cr
  [, 9] \tab women.job \tab\tab 1, 2, 3\cr
  [, 10] \tab money \tab\tab 1, 2, 3
  }
}
\source{
  K.Katsura and Y.Sakamoto (1980) \emph{Computer Science Monograph, No.14,
  CATDAP, A Categorical Data Analysis Program Package, DATA No.1.}
  The Institute of Statistical Mathematics.

  Y.Sakamoto, M.Ishiguro and G.Kitagawa (1983)
  \emph{Information Statistics}, III-2, DATA No. 9, Kyoritsu Shuppan Co., Ltd.,
  Tokyo. (in Japanese)
}

\keyword{datasets}
